\name{ibutt}
\alias{ibutt}
\docType{package}
\title{
\packageTitle{ibutt}
}
\description{
\packageDescription{A package for processing temperature and humidy data acquired from iButtons and Hydrochrons.}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{anRpackage}
\packageIndices{anRpackage}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{anRpackage}

Maintainer: \packageMaintainer{Daniel Noble}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
