% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{humidity}
\alias{humidity}
\title{Humidity data
Example humidity data taken from a hydrochron iButton}
\format{A raw csv file saved from the Hydrochron iButton}
\description{
Humidity data
Example humidity data taken from a hydrochron iButton
}

