% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibuttProc.R
\name{ibuttProc}
\alias{ibuttProc}
\title{Import and subset iButton files and data.}
\usage{
ibuttProc(dir, date, time, list = TRUE)
}
\arguments{
\item{dir}{The directory of the ibutton and/or hydrochron files.}

\item{date}{The date, as a character string, that one wishes to subset. Note that the format of the date should be "year-month-day" with only a four digit year.}

\item{time}{The hour (e.g. "11" or "14") or time range [e.g. c("11:30", "11:45")] one wishes to subset. Note that time should be given in a 24 hr clock. Time range is structured as a character string vector consisting of the lower time followed by the upper time limits as: c(lower time, upper time). When sub setting only a single hour period, simply just use the hour one is interested in sub setting. You must use two digits to to query each hour (i.e. "09" NOT "9").}

\item{list}{A logical argument specifying whether the data frame returned should be kept as a list or amalgamated into a data frame that can be queried using column names.}
}
\value{
A data.frame or list of temperatures/humidity across the dates and times that data was collected.
}
\description{
Import and subset iButton files and data.
}
\examples{
data <- ibuttProc("data/", date = "2015-11-02", time = c("07:00", "08:00"), list = TRUE)
}

