% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.func.R
\name{import}
\alias{import}
\title{Import a list of .csv files generated from iButtons and Hydrochrons. Save these files to a list and name the data.frames the file names}
\usage{
import(dir)
}
\arguments{
\item{dir}{Directory path to the folder containing the temperature/humidity data. Files should be .csv.}
}
\description{
Import a list of .csv files generated from iButtons and Hydrochrons. Save these files to a list and name the data.frames the file names
}

