% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibuttProc.R
\name{subtime}
\alias{subtime}
\title{Function for sub setting a specific hour or range of times}
\usage{
subtime(d, time)
}
\arguments{
\item{d}{The processed iButton data frame.}

\item{time}{The hour (e.g. "11" or "14") or time range (e.g. c("11:30", "11:45")) one wishes to subset. Note that time should be given in a 24 hr clock. Time range is structured as a character string vector consisting of the lower time followed by the upper time limits as: c(lower time, upper time). When sub setting only a single hour period, simply just use the hour one is interested in sub setting.}
}
\value{
A data.frame consisting of the sub temperature and humidity values that are subset by the hour or time range of interest.
}
\description{
Function for sub setting a specific hour or range of times
}
\examples{
#Get a range of temperatures
time <- subtime("./data/temp.csv", time = c("07:00", "08:00"))
#Get an hour
time <- subtime("./data/temp.csv", time = c("11"))
}

