% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibuttProc.R
\name{ibuttProc}
\alias{ibuttProc}
\title{Main function: Process and store ibutton data as lists or amalgamated data frame, which can be easily queried.}
\usage{
ibuttProc(dir, list = TRUE)
}
\arguments{
\item{dir}{The directory of the ibutton and/or hydrochron files.}

\item{list}{A logical argument specifying whether the data frame returned should be kept as a list or amalgamated into a data frame that can be queried using column names.}
}
\value{
A data.frame or list of temperatures/humidity across the dates and times that data was collected.
}
\description{
Main function: Process and store ibutton data as lists or amalgamated data frame, which can be easily queried.
}
\examples{
data <- ibuttProc("data/", list = TRUE)	
}

