% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process}
\alias{process}
\title{Function for processing the text read from ibuttons. Still needs work.}
\usage{
process(dat)
}
\arguments{
\item{dat}{a text/csv string imported from a directory containing csv files for ibutton data.}
}
\description{
Function for processing the text read from ibuttons. Still needs work.
}

